+++ R/create.R
@@ -1,8 +1,8 @@
-#' Compress Files into 'zip' Archives
+#' Compress files into ZIP archives
 #'
-#' `zip_create()` creates a new zip archive file.
-#'
-#' `zip_append()` appends compressed files to an existing 'zip' file.
+#' Compared to [zip::zip()] and [utils::zip()], [zip_create()] takes the path of
+#' files to include in the archive as the first argument. Append new compressed
+#' files to an existing archive with [zip_append()].
 #'
 #' ## Relative paths
 #'
@@ -25,111 +25,80 @@
 #'
 #' ## Permissions:
 #'
-#' `zip_create()` (and `zip_append()`, etc.) add the permissions of the archived
-#' files and directories to the ZIP archive, on Unix systems. Most zip and unzip
+#' `zip_create()` (and `zip_append()`) add the permissions of the archived files
+#' and directories to the ZIP archive, on Unix systems. Most zip and unzip
 #' implementations support these, so they will be recovered after extracting the
-#' archive.
+#' archive. Note, however that the owner and group (uid and gid) are currently
+#' omitted, even on Unix.
 #'
-#' Note, however that the owner and group (uid and gid) are currently omitted,
-#' even on Unix.
-#'
-#' ## `zipr()` and `zipr_append()`
-#'
-#' These function exist for historical reasons. They are identical to
-#' `zip_create()` and `zipr_append()` with a different default for the `mode`
-#' argument.
-#'
-#' @param zipfile The zip file to create. If the file exists, `zip` overwrites
-#'   it, but `zip_append` appends to it.
-#' @param files List of file to add to the archive. See details below about
-#'   absolute and relative path names.
+#' @param path Character vector of files to add to the archive.
+#' @param archive The ZIP file path to create. If the file exists,
+#'   [zip_create()] overwrites it, but [zip_append()] appends to it.
 #' @param recurse Whether to add the contents of directories recursively.
-#' @param compression_level A number between 1 and 9. 9 compresses best, but it
-#'   also takes the longest.
-#' @param include_directories Whether to explicitly include directories in the
-#'   archive. Including directories might confuse MS Office when reading docx
-#'   files, so set this to `FALSE` for creating them.
-#' @param root Change to this working directory before creating the archive.
-#' @param mode Selects how files and directories are stored in the archice. It
-#'   can be `"mirror"` or `"cherry-pick"`. See "Relative Paths" below for
-#'   details.
+#' @param level Integer between 1 and 9. Higher compression takes longer.
+#' @param junk_paths Whether to remove full file path. See "Relative paths"
+#'   below for more. An alternative to the `mode` argument of [zip::zip()].
+#' @param include_dirs Whether to explicitly include directories in the archive.
+#' @param root Directory to temporarily change to before creating archive.
 #'
 #' @examples
-#' ## Some files to zip up. We will run all this in the R sesion's
-#' ## temporary directory, to avoid messing up the user's workspace.
-#' dir.create(tmp <- tempfile())
-#' dir.create(file.path(tmp, "mydir"))
-#' cat("first file", file = file.path(tmp, "mydir", "file1"))
-#' cat("second file", file = file.path(tmp, "mydir", "file2"))
+#' \dontshow{.old_wd <- setwd(tempdir())}
+#' library(fs)
+#' dir_create(path("mydir"))
+#' cat("first file", file = path("mydir", "file1"))
+#' cat("second file", file = path("mydir", "file2"))
 #'
-#' zipfile <- tempfile(fileext = ".zip")
-#' zip_create(zipfile, "mydir", root = tmp)
+#' z <- file_temp(ext = "zip")
+#' zip_create("mydir", z)
 #'
 #' ## List contents
-#' zip_list(zipfile)
+#' zip_list(z)
 #'
 #' ## Add another file
-#' cat("third file", file = file.path(tmp, "mydir", "file3"))
-#' zip_append(zipfile, file.path("mydir", "file3"), root = tmp)
-#' zip_list(zipfile)
+#' cat("third file", file = path("mydir", "file3"))
+#' path("mydir", "file3") %>% zip_append(z)
+#' zip_list(z)
+#' \dontshow{setwd(.old_wd)}
 #'
-#' @return The name of the created zip file, invisibly.
+#' @return The name of the created ZIP archive, invisibly.
+#' @importFrom fs fs_path
 #' @export
-zip_create <- function(zipfile, files, recurse = TRUE, compression_level = 9,
-                       include_directories = TRUE, root = ".",
-                       mode = c("mirror", "cherry-pick")) {
-  mode <- match.arg(mode)
-  zip_internal(zipfile, files, recurse, compression_level, append = FALSE,
-               root = root, keep_path = (mode == "mirror"),
-               include_directories = include_directories)
+zip_create <- function(path, archive, recurse = TRUE, level = 9,
+                       junk_paths = FALSE, include_dirs = TRUE, root = ".") {
+  zip_internal(path, archive, recurse, level, append = FALSE,
+               root = root, junk_paths = junk_paths,
+               include_dirs = include_dirs)
 }
 
 #' @rdname zip_create
 #' @export
-zip_create2 <- function(zipfile, files, recurse = TRUE, compression_level = 9,
-                        include_directories = TRUE, root = ".",
-                        mode = c("cherry-pick", "mirror")) {
-  mode <- match.arg(mode)
-  zip_internal(zipfile, files, recurse, compression_level, append = FALSE,
-               root = root, keep_path = (mode == "mirror"),
-               include_directories = include_directories)
+zip_append <- function(path, archive, recurse = TRUE, level = 9,
+                       junk_paths = FALSE, include_dirs = TRUE, root = ".") {
+  zip_internal(path, archive, recurse, level, append = TRUE,
+               root = root, junk_paths = junk_paths,
+               include_dirs = include_dirs)
 }
 
-#' @rdname zip_create
-#' @export
-zip_append <- function(zipfile, files, recurse = TRUE,
-                       compression_level = 9, include_directories = TRUE,
-                       root = ".", mode = c("mirror", "cherry-pick")) {
-  mode <- match.arg(mode)
-  zip_internal(zipfile, files, recurse, compression_level, append = TRUE,
-               root = root, keep_path = (mode == "mirror"),
-               include_directories = include_directories)
-}
-
-#' @rdname zip_create
-#' @export
-zip_append2 <- function(zipfile, files, recurse = TRUE,
-                        compression_level = 9, include_directories = TRUE,
-                        root = ".", mode = c("cherry-pick", "mirror")) {
-  mode <- match.arg(mode)
-  zip_internal(zipfile, files, recurse, compression_level, append = TRUE,
-               root = root, keep_path = (mode == "mirror"),
-               include_directories = include_directories)
-}
-
-zip_internal <- function(zipfile, files, recurse, compression_level,
-                         append, root, keep_path, include_directories) {
+zip_internal <- function(path, archive, recurse, level, append, root,
+                         junk_paths, include_dirs) {
   oldwd <- setwd(root)
   on.exit(setwd(oldwd), add = TRUE)
 
-  if (any(! file.exists(files))) stop("Some files do not exist")
+  if (any(!file.exists(path))) {
+    stop("Some files do not exist", call. = FALSE)
+  }
 
-  data <- get_zip_data(files, recurse, keep_path, include_directories)
+  data <- get_zip_data(
+    files = path,
+    recurse = recurse,
+    keep_path = !junk_paths,
+    include_directories = include_dirs
+  )
   warn_for_dotdot(data$key)
 
-  .Call(c_R_zip_zip, enc2utf8(zipfile), enc2utf8(data$key),
+  .Call(c_R_zip_zip, enc2utf8(archive), enc2utf8(data$key),
         enc2utf8(data$file), data$dir, file.info(data$file)$mtime,
-        as.integer(compression_level), append)
+        as.integer(level), append)
 
-  invisible(zipfile)
+  invisible(fs::fs_path(archive))
 }